pyproject.toml
[tool.poetry]
name = "pixie-prompts-examples"
packages = [
    { include = "examples" }
]

version = "0.0.0"  # Managed by setuptools-scm
description = "examples for using pixie-prompts"
authors = ["Yiou Li <yol@gopixie.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.urls]
Homepage = "https://gopixie.ai"
Repository = "https://github.com/yiouli/pixie-prompts"
Documentation = "https://yiouli.github.io/pixie-prompts/"
Issues = "https://github.com/yiouli/pixie-prompts/issues"
Changelog = "https://github.com/yiouli/pixie-prompts/commits/main/"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pixie-prompts = {path = "../pixie-prompts", develop = true, extras = ["server"]}

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools-scm>=8.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools_scm]
write_to = "examples/_version.py"
